# python/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "allowed_decisions": {
            "items": {
              "enum": [
                "approve",
                "reject"
              ],
              "type": "string"
            },
            "title": "Allowed Decisions",
            "type": "array"
          },
          "approval_id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
            "title": "Approval Id",
            "type": "string"
          },
          "kind": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Kind",
            "type": "string"
          },
          "prompt": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "approval_id",
          "kind",
          "prompt",
          "allowed_decisions"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "CreateRenderJobRequest": {
        "additionalProperties": false,
        "description": "A render request naming one exact saved revision and nothing else.",
        "properties": {
          "document_id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
            "title": "Document Id",
            "type": "string"
          },
          "document_revision": {
            "exclusiveMinimum": 0.0,
            "title": "Document Revision",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "document_revision"
        ],
        "title": "CreateRenderJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RenderCapability": {
        "additionalProperties": false,
        "description": "Whether this installation can start a render right now, and why not.",
        "properties": {
          "active_render_job_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Render Job Id"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "preset_id": {
            "const": "standard_vertical_mp4_v1",
            "title": "Preset Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "renderer_not_configured",
                  "render_busy"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "renderer_version": {
            "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$",
            "title": "Renderer Version",
            "type": "string"
          }
        },
        "required": [
          "available",
          "preset_id",
          "renderer_version"
        ],
        "title": "RenderCapability",
        "type": "object"
      },
      "RenderJobPageView": {
        "additionalProperties": false,
        "description": "One bounded newest-first page of a project's render history.",
        "properties": {
          "jobs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RenderJobView"
            },
            "maxItems": 50,
            "title": "Jobs",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "RenderJobPageView",
        "type": "object"
      },
...
      "WorkflowFailure": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Code",
            "type": "string"
          },
          "failed_stage": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_]{0,63}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed Stage"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "WorkflowFailure",
        "type": "object"
      },
...
      "WorkflowSummary": {
        "additionalProperties": false,
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRef"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "$ref": "#/components/schemas/WorkflowSource"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageSummary"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_id": {
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "status",
          "created_at",
          "updated_at",
          "source",
          "stages"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Thoth Control Plane",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/legacy/jobs/{legacy_job_id}": {
      "get": {
        "description": "Return a safe v1 projection of one legacy Rust job, without mutations.",
        "operationId": "get_legacy_job_api_v1_legacy_jobs__legacy_job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "legacy_job_id",
            "required": true,
            "schema": {
              "title": "Legacy Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Observe a legacy job during migration",
        "tags": [
          "migration"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/render-capability": {
      "get": {
        "description": "Report whether this installation can start a render for this project now.",
        "operationId": "read_render_capability_api_v1_projects__project_id__render_capability_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderCapability"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Render Capability"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs": {
      "get": {
        "operationId": "list_render_jobs_api_v1_projects__project_id__render_jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobPageView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Render Jobs"
      },
      "post": {
        "description": "Start exactly one render of one saved revision, or report why not.",
        "operationId": "create_render_job_api_v1_projects__project_id__render_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRenderJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Render Job"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs/{render_job_id}": {
      "get": {
        "operationId": "read_render_job_api_v1_projects__project_id__render_jobs__render_job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "render_job_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
              "title": "Render Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Render Job"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs/{render_job_id}/artifacts": {
      "delete": {
        "description": "Delete one terminal job's files by hand, keeping its audit row.",
        "operationId": "cleanup_render_artifacts_api_v1_projects__project_id__render_jobs__render_job_id__artifacts_delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "render_job_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
              "title": "Render Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup Render Artifacts"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs/{render_job_id}/cancel": {
      "post": {
        "description": "Record one cancel request; the renderer answers through its own events.",
        "operationId": "cancel_render_job_api_v1_projects__project_id__render_jobs__render_job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "render_job_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
              "title": "Render Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Render Job"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs/{render_job_id}/output": {
      "get": {
        "description": "Stream the one published render, never a redirect to where it lives.",
        "operationId": "download_render_output_api_v1_projects__project_id__render_jobs__render_job_id__output_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "render_job_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
              "title": "Render Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "video/mp4": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Render Output"
      }
    },
    "/api/v1/projects/{project_id}/render-jobs/{render_job_id}/retry": {
      "post": {
        "description": "Start a new job for the same revision; the finished one stays finished.",
        "operationId": "retry_render_job_api_v1_projects__project_id__render_jobs__render_job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "pattern": "^(?:[A-Za-z][A-Za-z0-9_-]{0,127}|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "render_job_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,127}$",
              "title": "Render Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Render Job"
      }
    },
...
  }
}